Remove contrib neon and neon_test_utils.
