--- contrib/neon/neon--1.0.sql
@@ -1,17 +1,0 @@
-\echo Use "CREATE EXTENSION neon" to load this file. \quit
-
-CREATE FUNCTION pg_cluster_size()
-RETURNS bigint
-AS 'MODULE_PATHNAME', 'pg_cluster_size'
-LANGUAGE C STRICT
-PARALLEL UNSAFE;
-
-CREATE FUNCTION backpressure_lsns(
-    OUT received_lsn pg_lsn,
-    OUT disk_consistent_lsn pg_lsn,
-    OUT remote_consistent_lsn pg_lsn
-)
-RETURNS record
-AS 'MODULE_PATHNAME', 'backpressure_lsns'
-LANGUAGE C STRICT
-PARALLEL UNSAFE;
