--- contrib/neon/neon.control
@@ -1,4 +1,0 @@
-# neon extension
-comment = 'cloud storage for PostgreSQL'
-default_version = '1.0'
-module_pathname = '$libdir/neon'
