--- contrib/neon_test_utils/neon_test_utils.control
@@ -1,5 +1,0 @@
-# neon_test_utils extension
-comment = 'helpers for neon testing and debugging'
-default_version = '1.0'
-module_pathname = '$libdir/neon_test_utils'
-relocatable = true
