+++ src/backend/main/main.c
@@ -39,7 +39,7 @@
 #include "common/username.h"
 #include "port/atomics.h"
 #include "postmaster/postmaster.h"
-#include "replication/walproposer.h"
+#include "replication/walpropshim.h"
 #include "storage/spin.h"
 #include "tcop/tcopprot.h"
 #include "utils/help_config.h"
