--- src/backend/replication/libpqwalproposer/Makefile
@@ -1,37 +1,0 @@
-#-------------------------------------------------------------------------
-#
-# Makefile--
-#    Makefile for src/backend/replication/libpqwalproposer
-#
-# IDENTIFICATION
-#    src/backend/replication/libpqwalproposer/Makefile
-#
-#-------------------------------------------------------------------------
-
-subdir = src/backend/replication/libpqwalproposer
-top_builddir = ../../../..
-include $(top_builddir)/src/Makefile.global
-
-override CPPFLAGS := -I$(srcdir) -I$(libpq_srcdir) $(CPPFLAGS)
-
-OBJS = \
-	$(WIN32RES) \
-	libpqwalproposer.o
-SHLIB_LINK_INTERNAL = $(libpq)
-SHLIB_LINK = $(filter -lintl, $(LIBS))
-SHLIB_PREREQS = submake-libpq
-PGFILEDESC = "libpqwalproposer - libpq interface for WAL proposer"
-NAME = libpqwalproposer
-
-all: all-shared-lib
-
-include $(top_srcdir)/src/Makefile.shlib
-
-install: all installdirs install-lib
-
-installdirs: installdirs-lib
-
-uninstall: uninstall-lib
-
-clean distclean maintainer-clean: clean-lib
-	rm -f $(OBJS)
