+++ src/include/replication/walsender.h
@@ -48,8 +48,25 @@ extern void WalSndInitStopping(void);
 extern void WalSndWaitStopping(void);
 extern void HandleWalSndInitStopping(void);
 extern void WalSndRqstFileReload(void);
-extern void GetMinReplicaLsn(XLogRecPtr* write, XLogRecPtr* flush, XLogRecPtr* apply);
-extern uint64 backpressure_lag(void);
+
+/*
+ * Hook to check for WAL receiving backpressure.
+ * Return value in microseconds */
+extern uint64 (*delay_backend_us)(void);
+
+/* expose these so that they can be reused by the neon walproposer extension */
+extern void LagTrackerWrite(XLogRecPtr lsn, TimestampTz local_flush_time);
+extern TimeOffset LagTrackerRead(int head, XLogRecPtr lsn, TimestampTz now);
+extern void ProcessStandbyReply(XLogRecPtr writePtr, XLogRecPtr flushPtr,
+								XLogRecPtr applyPtr, TimestampTz replyTime,
+								bool replyRequested);
+void       PhysicalConfirmReceivedLocation(XLogRecPtr lsn);
+void       ProcessStandbyHSFeedback(TimestampTz   replyTime,
+									TransactionId feedbackXmin,
+									uint32		feedbackEpoch,
+									TransactionId feedbackCatalogXmin,
+									uint32		feedbackCatalogEpoch);
+
 /*
  * Remember that we want to wakeup walsenders later
  *
